feat(es/transform): introduce `NodeIgnoringSpan`
