+++ crates/swc_ecma_utils/Cargo.toml
@@ -1,16 +1,16 @@
 [package]
-authors = ["강동윤 <kdy1997.dev@gmail.com>"]
-description = "Utilities for swc ecmascript ast nodes"
+authors       = ["강동윤 <kdy1997.dev@gmail.com>"]
+description   = "Utilities for swc ecmascript ast nodes"
 documentation = "https://rustdoc.swc.rs/swc_ecma_utils/"
-edition = "2021"
-license = "Apache-2.0"
-name = "swc_ecma_utils"
-repository = "https://github.com/swc-project/swc.git"
-version = "0.111.4"
+edition       = "2021"
+license       = "Apache-2.0"
+name          = "swc_ecma_utils"
+repository    = "https://github.com/swc-project/swc.git"
+version       = "0.111.4"
 
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
+    [package.metadata.docs.rs]
+    all-features = true
+    rustdoc-args = ["--cfg", "docsrs"]
 
 [lib]
 bench = false
@@ -20,17 +20,18 @@ bench = false
 concurrent = ["swc_common/concurrent", "rayon"]
 
 [dependencies]
-indexmap = "1.6.1"
-num_cpus = "1.13.1"
-once_cell = "1.10.0"
-rayon = {version = "1.5.1", optional = true}
-rustc-hash = "1.1.0"
-swc_atoms = {version = "0.4.39", path = "../swc_atoms"}
-swc_common = {version = "0.29.35", path = "../swc_common"}
-swc_ecma_ast = {version = "0.98.2", path = "../swc_ecma_ast"}
-swc_ecma_visit = {version = "0.84.2", path = "../swc_ecma_visit"}
-tracing = "0.1.32"
-unicode-id = "0.3"
+indexmap       = "1.6.1"
+num_cpus       = "1.13.1"
+once_cell      = "1.10.0"
+rayon          = { version = "1.5.1", optional = true }
+rustc-hash     = "1.1.0"
+swc_atoms      = { version = "0.4.39", path = "../swc_atoms" }
+swc_common     = { version = "0.29.35", path = "../swc_common" }
+swc_ecma_ast   = { version = "0.98.2", path = "../swc_ecma_ast" }
+swc_ecma_visit = { version = "0.84.2", path = "../swc_ecma_visit" }
+tracing        = "0.1.32"
+unicode-id     = "0.3"
 
 [dev-dependencies]
-swc_ecma_parser = {version = "0.128.4", path = "../swc_ecma_parser"}
+swc_ecma_parser = { version = "0.128.4", path = "../swc_ecma_parser" }
+testing         = { version = "0.31.37", path = "../testing" }
